Move Fs trait into its own module
